Closes #32; fully separated extensions from core.
