+++ app/core/app_template/urls.py-tpl
@@ -1,5 +1,5 @@
 from django.urls import path
-from core.utilities.types import URLPatternsList
+from extensions.utilities.types import URLPatternsList
 from {{ app_name }} import views
