+++ app/core/exceptions/formatter.py
@@ -1,7 +1,7 @@
 from dataclasses import asdict
 from drf_standardized_errors.formatter import ExceptionFormatter as BaseExceptionFormatter
 from drf_standardized_errors.types import ErrorResponse
-from core.utilities.types import JSON
+from extensions.utilities.types import JSON
 
 
 class ExceptionFormatter(BaseExceptionFormatter):
