+++ app/users/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from core.utilities.types import URLPatternsList
+from extensions.utilities.types import URLPatternsList
 from users import views
